Updated Allegro postman collection (Groups, Group Members)
When using this postman group, add/use a Postman environment.
This will allow environment variables to be set, which are
used throughout the URLs in Groups. For instance, a new group
can be created and there's no need to copy/paste the group ID
when trying out the "Show group details" endpoint. Postman
does it for you. Isn't that great? :)