Added new identity config option: "interface"
Valid values are "admin" or "public" (default is admin).
Use 'public' if the admin interface is unreachable. However,
this means tenant visibility to allegro-api will be limited
to tenants that the allegro user is at least a member of. This
is due to how keystone handles requests across interfaces. The
admin interface must be used to see all tenants, provided allegro
has an admin role in at least one tenant (usually 'service').