Revised Postman collection.
The Postman environment must now have these variables at a minimum:

  tenant_id:TENANT_ID_FOR_URL
  allegro:http://allegro:8090
  keystone:http://qos101.research.att.com:5000
  username:allegro
  password:ALLEGRO_PASSWORD
  tenant_name:service

Make sure the allegro user has been added to the service tenant
as an admin.

The tenant_id is automatically injected into all allegro-api
Postman URLs. This must be a tenant_id that is "visible" by the
allegro user.

allegro and keystone are publicurl endpoints for each (no trailing
slash). These urls should *always* be the publicurl versions.

Keystone may opt to use the adminurl endpoint once it authenticates
a token and executes a particular command, but this can be
overridden with caveats.