Placements endpoint now supports POST for possible replanning.
The request body is expected to contain a list of candidate
locations (typically obtained from nova-scheduler).

Let's say nova-scheduler has these as available locations for
a particular VM:

{
   "locations": ["qos101", "qos104", "qos106", "qos107"]
}

If Ostro has already planned that VM, and the location is in
this list, it will be returned as usual. If the location is NOT
in this list, Ostro will be asked to replan, given these as
possible backup locations. In the latter case, Ostro will replan
at LEAST this VM, and possibly more that haven't already been
scheduled by nova-scheduler. In any event, the updated placement
for this VM will then be returned.