Nova ValetFilter now supports ad-hoc placements.
If a resource is being scheduled and the request did not
come from heat (no stack lifecycle scheduler hints), place
the resource on-the-fly. The stack id and orchestration id
will be set to the physical resource id in this case, and
an ad hoc resource will be sent to valet, containing the
flavor and no other properties.