Added a failure_mode option under [valet] for nova.conf.
Valid options are 'reject' and 'yield'. Default is 'reject'.

Set to 'yield' if nova ValetFilter is to yield to Nova for placement
in the event that Valet/Ostro/etc. can't be reached or otherwise
can't decide on a placement. Note that this can risk violating
exclusivity requirements if 'yield' is chosen.