Covering for Ostro when there are no serviceable resources.
Ostro is currently throwing an error if it is asked to solve
placements with no serviceable resources. Ostro should succeed
and return an empty placement list. For now, valet-api, upon
plan creation, will skip calling Ostro if no serviceable
resources are in the request. It will not fail but succeed
with a 201 Status and an empty plan object.